add simple support for eclipse projects
For maven and eclipse projects, the classpath will contain JAVA_HOME rather than
a configured java toolchain. If JAVA_HOME is not set, classes from the standard
library will not be found. However, that does not prevent the tool from finding
all classes that are outside of the standard library.